Moved placeholders to eCloud
