+++ .idea/compiler.xml
@@ -7,6 +7,7 @@
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
         <module name="tmoney-spigot" />
+        <module name="tmoney-papi" />
         <module name="tmoney-api" />
         <module name="tmoney-common" />
       </profile>
