+++ tmoney-spigot/dependency-reduced-pom.xml
@@ -3,7 +3,7 @@
   <parent>
     <artifactId>TMoney</artifactId>
     <groupId>me.theseems.tmoney</groupId>
-    <version>0.1D</version>
+    <version>0.2D</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <artifactId>tmoney-spigot</artifactId>
@@ -69,72 +69,72 @@
         </exclusion>
         <exclusion>
           <artifactId>snakeyaml</artifactId>
-            <groupId>org.yaml</groupId>
+          <groupId>org.yaml</groupId>
         </exclusion>
       </exclusions>
     </dependency>
-      <dependency>
-          <groupId>me.clip</groupId>
-          <artifactId>placeholderapi</artifactId>
-          <version>2.9.2</version>
-          <scope>provided</scope>
-      </dependency>
-      <dependency>
-          <groupId>com.github.MilkBowl</groupId>
-          <artifactId>VaultAPI</artifactId>
-          <version>1.7</version>
-          <scope>provided</scope>
-          <exclusions>
-              <exclusion>
-                  <artifactId>bukkit</artifactId>
-                  <groupId>org.bukkit</groupId>
-              </exclusion>
-          </exclusions>
-      </dependency>
-      <dependency>
-          <groupId>org.postgresql</groupId>
-          <artifactId>postgresql</artifactId>
-          <version>42.2.5</version>
-          <scope>provided</scope>
-      </dependency>
-      <dependency>
-          <groupId>com.h2database</groupId>
-          <artifactId>h2</artifactId>
-          <version>1.4.200</version>
-          <scope>provided</scope>
-      </dependency>
-      <dependency>
-          <groupId>junit</groupId>
-          <artifactId>junit</artifactId>
-          <version>RELEASE</version>
-          <scope>test</scope>
-          <exclusions>
-              <exclusion>
-                  <artifactId>hamcrest-core</artifactId>
-                  <groupId>org.hamcrest</groupId>
-              </exclusion>
-          </exclusions>
-      </dependency>
-      <dependency>
+    <dependency>
+      <groupId>me.clip</groupId>
+      <artifactId>placeholderapi</artifactId>
+      <version>2.9.2</version>
+      <scope>provided</scope>
+    </dependency>
+    <dependency>
+      <groupId>com.github.MilkBowl</groupId>
+      <artifactId>VaultAPI</artifactId>
+      <version>1.7</version>
+      <scope>provided</scope>
+      <exclusions>
+        <exclusion>
+          <artifactId>bukkit</artifactId>
+          <groupId>org.bukkit</groupId>
+        </exclusion>
+      </exclusions>
+    </dependency>
+    <dependency>
+      <groupId>org.postgresql</groupId>
+      <artifactId>postgresql</artifactId>
+      <version>42.2.5</version>
+      <scope>provided</scope>
+    </dependency>
+    <dependency>
+      <groupId>com.h2database</groupId>
+      <artifactId>h2</artifactId>
+      <version>1.4.200</version>
+      <scope>provided</scope>
+    </dependency>
+    <dependency>
+      <groupId>junit</groupId>
+      <artifactId>junit</artifactId>
+      <version>RELEASE</version>
+      <scope>test</scope>
+      <exclusions>
+        <exclusion>
+          <artifactId>hamcrest-core</artifactId>
+          <groupId>org.hamcrest</groupId>
+        </exclusion>
+      </exclusions>
+    </dependency>
+    <dependency>
+      <groupId>org.junit.jupiter</groupId>
+      <artifactId>junit-jupiter</artifactId>
+      <version>RELEASE</version>
+      <scope>test</scope>
+      <exclusions>
+        <exclusion>
+          <artifactId>junit-jupiter-api</artifactId>
           <groupId>org.junit.jupiter</groupId>
-          <artifactId>junit-jupiter</artifactId>
-          <version>RELEASE</version>
-          <scope>test</scope>
-          <exclusions>
-              <exclusion>
-                  <artifactId>junit-jupiter-api</artifactId>
-                  <groupId>org.junit.jupiter</groupId>
-              </exclusion>
-              <exclusion>
-                  <artifactId>junit-jupiter-params</artifactId>
-                  <groupId>org.junit.jupiter</groupId>
-              </exclusion>
-              <exclusion>
-                  <artifactId>junit-jupiter-engine</artifactId>
-                  <groupId>org.junit.jupiter</groupId>
-              </exclusion>
-          </exclusions>
-      </dependency>
+        </exclusion>
+        <exclusion>
+          <artifactId>junit-jupiter-params</artifactId>
+          <groupId>org.junit.jupiter</groupId>
+        </exclusion>
+        <exclusion>
+          <artifactId>junit-jupiter-engine</artifactId>
+          <groupId>org.junit.jupiter</groupId>
+        </exclusion>
+      </exclusions>
+    </dependency>
   </dependencies>
 </project>
